backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
psycopg2-binary
web3==6.11.3
eth-account==0.10.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.1.2
gunicorn==21.2.0
eventlet==0.35.1
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
cryptography==41.0.7
pyjwt==2.8.0
marshmallow==3.20.1
pytest==7.4.3
python-dateutil==2.8.2
pyotp==2.9.0
qrcode[pil]==7.4.2

# TimescaleDB support (works with PostgreSQL)
# No additional package needed - TimescaleDB is a PostgreSQL extension
# psycopg2-binary handles the connection